// web/src/api/generated/model/transferResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * MyFinances.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface TransferResponse {
  id: string
  fromAccountId: string
  toAccountId: string
  amount: number
  /** @nullable */
  note: string | null
  bookedOn: string
  createdAt: string
}